tests: split display-formatter tests into control-logic-display.test.js
tests/control-logic.test.js grew to 1243 lines after the mean-tank /
stall-bypass tests were added, tripping the 1200-line hard cap on test
files that check:file-size --strict enforces in CI.

Extracted the three display-formatter describe blocks (formatDuration,
formatTemp, buildDisplayLabels) into tests/control-logic-display.test.js
— they are pure helpers that don't touch evaluate() or makeState(), so
they make a natural split. The remaining control-logic.test.js now
exercises only the evaluator state machine and device-config gating.

Sizes after split:
  tests/control-logic.test.js          1123 lines  (was 1243)
  tests/control-logic-display.test.js   130 lines  (new)

0 files over hard cap; 18 files over soft cap (same 18 as main).
808/808 unit tests still pass.
